Modernize MCP stack, keep legacy behind --enable-legacy-mcp flag
(cherry picked from commit 120868a65438a7f8eb4f20b37cdd53dfd982f4f0)

+++ src/Spice86.Core/Emulator/Mcp/Response/IoPortReadResponse.cs
@@ -8,7 +8,7 @@ public sealed record IoPortReadResponse : McpToolResponse {
     /// <summary>
     /// The IO port number that was read.
     /// </summary>
-    public required ushort Port { get; init; }
+    public required int Port { get; init; }
     /// <summary>
     /// The value read from the IO port.
     /// </summary>
